refactor: move utils
